1) Eliminate static methods from HGDBApplication, OWLDataFactoryHGDB and some other related classes. Those should not be singletons because there is no reason to restrict 1 graph database instance per JVM for OWL. Also, the choice of whether this graph is versioned or distributed etc. was compile time, defined in a static variable, so that was removed as well.
2) Get GarbageCollector instance out of HGDBOntologyRepository class. It has no use there. 
The collector can be instantiated and run on a need by need basis.

3) Eliminate static getIntance methods from (VD)HGDBOntologyRepository. For that we need to be
able to create relatively lightweight instances of repositories, ontology managers, factories
attached to a currently open HG instance in the environment. That is because there is caching
going on and other RAM data structures that need to be in sync with the database. Every static
method that produces a manager/global kind of object of the OWLAPI will take a graph location
parameter and return the associated object with that graph. So the general problem here is
singleton objects in the context of a HyperGraph. So it's kind of like the singleton pattern,
but constrained to some fixed reference - in our case the absolute path/location of a database
instance. 

4) All in all, it is now possible to have multiple HGDB repositories open at the same time, and
one can obtain a plain, a versioned or a distributed versioned client/interface to interact
with them (basically HGDBOntologyRepository and derivatives are now lightweight objects that can
be created and destroyed at will). 

5) What remains a relatively big object is the HGDBOntologyManagerImpl which inherits from
the OWLAPI manager, which maintains tons of data structures, so this is not something you
want to create a lot of.